rknn_api/pyproject.toml
[tool.poetry]
name = "rknn-api"
version = "0.1.0"
description = ""
authors = ["Henrycke Bozza - Parallax <henrycke.bozza@orakolo.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
python-multipart = "^0.0.17"
rknn_toolkit2 = { file = "build/rknn_toolkit2-2.2.0-cp38-cp38-manylinux_2_17_x86_64.manylinux2014_x86_64.whl" }
protobuf = "3.20.3"
psutil = ">=5.9.0,<5.10.0"
torch = ">=2.0.0, !=2.0.1, !=2.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
